fix: extract CLI command registry, fix build config (#488)
- Extract 25 inline command definitions from cli.ts (826→62 lines) into
  cli-registry.ts with CLICommandDef interface and localOnly metadata
- Replace hardcoded LOCAL_ONLY_COMMANDS array with Set derived from
  registry, eliminating sync drift risk
- Downgrade @types/node from ^25 to ^20 to match engines.node >=20
- Add --sourcemap to esbuild bundle script for production debugging
- Set skipLibCheck: false, fix surfaced type error in cachedSearch
  (string→union types for sort/order params)
- Add snooze --days input validation consistent with search count
- Preserve lazy-loading pattern via dynamic import() in action handlers
